requirements.txt
feedparser~=6.0.11
python-slugify~=8.0.4
